backend/requirements.txt
# FastAPI framework
fastapi==0.95.2

# Database and ORM
sqlalchemy==1.4.47
psycopg2-binary==2.9.6

# Password hashing
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# JWT token generation
python-jose==3.3.0

# Background tasks with Celery
celery[redis]==5.3.0
redis==4.5.5

# Static file serving
aiofiles==22.1.0

# ASGI server for FastAPI
uvicorn==0.23.0

# Optional for handling .env files
python-dotenv==1.0.0
